# dist/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OZMAP API",
        "termsOfService": "http://localhost:8080/terms",
        "contact": {
            "email": "wevertonbezzera013@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/",
            "description": "Test API"
        }
    ],
    "paths": {
...
        "/export/regions": {
            "get": {
                "summary": "Export Regions to CSV",
                "description": "This route is used to export regions data to a CSV file.",
                "tags": ["export"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/export/users": {
            "get": {
                "summary": "Export Users to CSV",
                "description": "This route is used to export users data to a CSV file.",
                "tags": ["export"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
...
}